+++ examples/terraform/multi-root-green-path/06-ses/locals.tf
@@ -1,0 +1,3 @@
+locals {
+  template_name_otp = "${var.project_name}-otp-${var.environment}"
+}